requirements.txt
dotenv
fastapi
python-dotenv
python-telegram-bot
uvicorn